stream added, rename my product, delete some code in maven, changer constructor in classes...

+++ src/main/java/com/internet/shop/model/Product.java
@@ -1,13 +1,11 @@
 package com.internet.shop.model;
 
-import java.math.BigDecimal;
-
 public class Product {
     private Long id;
     private String name;
-    private BigDecimal price;
+    private double price;
 
-    public Product(String name, BigDecimal price) {
+    public Product(String name, double price) {
         this.name = name;
         this.price = price;
     }
@@ -20,7 +18,7 @@ public class Product {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
@@ -32,7 +30,7 @@ public class Product {
         this.name = name;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
